# docs/swagger.yaml
basePath: /
definitions:
...
  models.ReceiptRequest:
    properties:
      amount:
        example: 49.99
        type: number
      currency:
        example: NGN
        type: string
      description:
        example: Pro subscription payment
        type: string
      payer_email:
        example: user@example.com
        type: string
      payer_name:
        example: Jane Doe
        type: string
      payment_method:
        example: card
        type: string
      reference:
        example: PAY-20260821-001
        type: string
      status:
        example: success
        type: string
      transaction_date:
        example: "2026-08-21T08:30:00Z"
        type: string
    type: object
...
  routes.AdvanceScanRequest:
    properties:
      content:
        example: Claim your reward at https://secure-login-example.com/verify before
          it expires.
        type: string
    type: object
  routes.BasicScanRequest:
    properties:
      content:
        example: Your account will be suspended unless you confirm your password now.
        type: string
    type: object
...
  routes.MessageScanRequest:
    properties:
      content:
        example: Please send the 0.5 ETH to this address right away so we can release
          the funds.
        type: string
    type: object
  routes.MidScanRequest:
    properties:
      content:
        example: We detected unusual activity. Please update your payment details
          immediately.
        type: string
    type: object
...
  routes.ScanRequest:
    properties:
      content:
        example: Congratulations! You have won a prize. Click this link now.
        type: string
    type: object
...
  routes.URLScanRequest:
    properties:
      content:
        example: https://secure-login-example.com/account/verify?session=123
        type: string
    type: object
...
  routes.WalletScanRequest:
    properties:
      content:
        example: 0x3fD2b5e4c7A9E2d6C1b8F0a3D4e5F67890123456
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@ganium.ai
    name: Ganium Support
  description: Ganium AI scam detection, security investigation, scanning, authentication
    and payment API.
  license:
    name: Proprietary
  termsOfService: https://ganium.ai/terms
  title: Ganium API
  version: "1.0"
paths:
...
  /api/receipt/pdf:
    post:
      consumes:
      - application/json
      description: Generates a PDF payment receipt for a transaction.
      parameters:
      - description: Receipt information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ReceiptRequest'
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Generate payment receipt PDF
      tags:
      - PDF
  /api/scan/{scanType}/pdf:
    post:
      consumes:
      - application/json
      description: Generates a PDF report for a security scan.
      parameters:
      - description: Scan type
        enum:
        - basic
        - mid
        - advance
        - message
        - url
        - wallet
        in: path
        name: scanType
        required: true
        type: string
      - description: Content to scan for the selected scanType
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.ScanRequest'
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Generate scan PDF
      tags:
      - PDF
  /api/scan/advance:
    post:
      consumes:
      - application/json
      description: Performs an advanced security scan on supplied content.
      parameters:
      - description: Text content to scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.AdvanceScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Advanced security scan
      tags:
      - Scan
  /api/scan/basic:
    post:
      consumes:
      - application/json
      description: Performs a basic security scan on supplied content.
      parameters:
      - description: Text content to scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.BasicScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Basic security scan
      tags:
      - Scan
  /api/scan/message:
    post:
      consumes:
      - application/json
      description: Analyzes a message for scams, phishing, and malicious content.
      parameters:
      - description: Message content to scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.MessageScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Scan message
      tags:
      - Scan
  /api/scan/mid:
    post:
      consumes:
      - application/json
      description: Performs a mid-level security scan on supplied content.
      parameters:
      - description: Text content to scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.MidScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mid-level security scan
      tags:
      - Scan
  /api/scan/url:
    post:
      consumes:
      - application/json
      description: Analyzes a URL for malicious, phishing, and scam indicators.
      parameters:
      - description: URL to scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.URLScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Scan URL
      tags:
      - Scan
  /api/scan/wallet:
    post:
      consumes:
      - application/json
      description: Analyzes a cryptocurrency wallet address for suspicious activity.
      parameters:
      - description: Wallet address to scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.WalletScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Scan wallet
      tags:
      - Scan
...
securityDefinitions:
  BearerAuth:
    description: 'JWT authentication. Enter your token using the format: Bearer <token>.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"